+++ cmd/common.go
@@ -56,8 +56,8 @@ type CommonConfig struct {
 
 func AddCommonFlags(cmd *cobra.Command, cfg *CommonConfig) {
 	cmd.PersistentFlags().StringVar(&cfg.ClusterName, "cluster-name", "", "cluster name")
-	cmd.PersistentFlags().StringVar(&cfg.StoreBackend, "store-backend", "", "store backend type (etcdv2/etcd, etcdv3, consul or kubernetes)")
-	cmd.PersistentFlags().StringVar(&cfg.StoreEndpoints, "store-endpoints", "", "a comma-delimited list of store endpoints (use https scheme for tls communication) (defaults: http://127.0.0.1:2379 for etcd, http://127.0.0.1:8500 for consul)")
+	cmd.PersistentFlags().StringVar(&cfg.StoreBackend, "store-backend", "", "store backend type (etcdv3 or kubernetes)")
+	cmd.PersistentFlags().StringVar(&cfg.StoreEndpoints, "store-endpoints", "", "a comma-delimited list of store endpoints (use https scheme for tls communication) (defaults: http://127.0.0.1:2379 for etcdv3)")
 	cmd.PersistentFlags().DurationVar(&cfg.StoreTimeout, "store-timeout", cluster.DefaultStoreTimeout, "store request timeout")
 	cmd.PersistentFlags().StringVar(&cfg.StorePrefix, "store-prefix", common.StorePrefix, "the store base prefix")
 	cmd.PersistentFlags().StringVar(&cfg.StoreCertFile, "store-cert-file", "", "certificate file for client identification to the store")
@@ -107,11 +107,6 @@ func CheckCommonConfig(cfg *CommonConfig) error {
 	}
 
 	switch cfg.StoreBackend {
-	case "consul":
-	case "etcd":
-		// etcd is old alias for etcdv2
-		cfg.StoreBackend = "etcdv2"
-	case "etcdv2":
 	case "etcdv3":
 	case "kubernetes":
 		if cfg.KubeResourceKind == "" {
@@ -158,10 +153,6 @@ func NewStore(cfg *CommonConfig) (store.Store, error) {
 	var s store.Store
 
 	switch cfg.StoreBackend {
-	case "consul":
-		fallthrough
-	case "etcdv2":
-		fallthrough
 	case "etcdv3":
 		storePath := filepath.Join(cfg.StorePrefix, cfg.ClusterName)
 
@@ -188,10 +179,6 @@ func NewElection(cfg *CommonConfig, uid string) (store.Election, error) {
 	var election store.Election
 
 	switch cfg.StoreBackend {
-	case "consul":
-		fallthrough
-	case "etcdv2":
-		fallthrough
 	case "etcdv3":
 		storePath := filepath.Join(cfg.StorePrefix, cfg.ClusterName)
 
@@ -199,7 +186,10 @@ func NewElection(cfg *CommonConfig, uid string) (store.Election, error) {
 		if err != nil {
 			return nil, fmt.Errorf("cannot create kv store: %v", err)
 		}
-		election = store.NewKVBackedElection(kvstore, filepath.Join(storePath, common.SentinelLeaderKey), uid, cfg.StoreTimeout)
+		election, err = store.NewKVBackedElection(kvstore, filepath.Join(storePath, common.SentinelLeaderKey), uid, cfg.StoreTimeout)
+		if err != nil {
+			return nil, err
+		}
 	case "kubernetes":
 		kubecli, podName, namespace, err := getKubeValues(cfg)
 		if err != nil {